requirements.txt
streamlit==1.31.0
qrcode==7.4.2
Pillow==10.2.0
python-dateutil==2.8.2
